Abstracted out `ufmjs()` to its own Marked extension file
making it simpler to add unit-tests.
