+++ src/Umbraco.Web.UI.Client/src/packages/ufm/plugins/index.ts
@@ -1,1 +1,2 @@
 export { ufm } from './marked-ufm.plugin.js';
+export { ufmjs } from './marked-ufmjs.plugin.js';
